ci: build datum inside the CI
As this could be a potential spam point for our GitHub because
people might submit code without verifying if it compiles.

However, having this inside the CI allows us to cover a basic compilation
test, which for some contributions may not require a full build. Additionally,
we can ensure that the code compiles on basic Linux OS environments, like Ubuntu,
or detect if there are any dependency breaks (even with minimal dependencies,
so this shouldn't be an issue either).

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>