stratum: Add ability to map specific username suffixes to split shares
If a username contains a tilde character ('~'), the suffix following it is looked up in a mapping table specified in the configuration file.
The mapping entries specify alternate Bitcoin addresses to submit shares with, up to a specified proportion.
The worker name specified before the tilde following a period ('.'), if any, is copied to each mapped address.
Mappings may include a "" entry that assigns a split to the original username/address.
If less than 100% is assigned, the remaining goes to the address configured by mining.pool_address (without any worker name!).
If greater than 100% is assigned, only the first 100% (in random order!) will get shares distributed.